% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_articles_year.R
\name{get_articles_year}
\alias{get_articles_year}
\title{get_articles_year}
\format{
Year:
\describe{
  \item{x}{placeholder}

}
}
\usage{
get_articles_year(i)
}
\description{
Get a Single Years Worth of Articles
}
