% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_articles_year.R
\name{get_articles_year}
\alias{get_articles_year}
\title{get_articles_year}
\usage{
get_articles_year(year, ...)
}
\arguments{
\item{year}{A year}

\item{...}{other arguments passed to methods}
}
\description{
Get a Single Years Worth of Articles
}
\examples{
\dontrun{
get_articles_year("1977", ...)
}
}
