% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_articles.R
\name{search_articles}
\alias{search_articles}
\title{search_articles}
\usage{
search_articles(query = query, year = NULL)
}
\description{
Search by Keyword Across all Years or in One Specific Year
}
\examples{
search_articles(query = "police", year = "1977")
}
