% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_articles.R
\name{search_articles}
\alias{search_articles}
\title{search_articles}
\usage{
search_articles(query = query, year = NULL, strip_html = FALSE, ...)
}
\arguments{
\item{query}{A Keyword to run}

\item{year}{A Year}

\item{strip_html}{Removes <HTML> tags}

\item{...}{other arguments passed to methods}
}
\description{
Search by Keyword Across all Years or in One Specific Year
}
\examples{
\dontrun{
search_articles(query = "Brexit", year = "2016", strip_html = FALSE, ...)
}
}
